src/gcf/requirements.txt
functions-framework==3.*
google-cloud-bigquery
pandas
numpy
db-dtypes
requests
